사운드 작업 완료 //  글꼴,레벨테스트, 이펙트 필요

+++ NoTouch/NoTouch/Assets/Scrtipt/Gem.cs
@@ -14,7 +14,7 @@ public class Gem : MonoBehaviour
 
     private double[] mShiftGap;
     private double mGemCost;
-    private float mGemProgress;
+    private double mGemProgress;
     private int mCurrentImageIndex;
     private double GapCal;
